Restructure python_compiler
