+++ pennylane/compiler/python_compiler/transforms/quantum/__init__.py
@@ -11,33 +11,21 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""PennyLane-xDSL transformations API."""
-
-from .mbqc import (
-    convert_to_mbqc_formalism_pass,
-    ConvertToMBQCFormalismPass,
-    decompose_graph_state_pass,
-    DecomposeGraphStatePass,
-    null_decompose_graph_state_pass,
-    NullDecomposeGraphStatePass,
-)
-
-from .quantum import (
-    combine_global_phases_pass,
-    CombineGlobalPhasesPass,
-    diagonalize_final_measurements_pass,
+"""PennyLane-xDSL API for quantum transforms"""
+from .cancel_inverses import IterativeCancelInversesPass, iterative_cancel_inverses_pass
+from .combine_global_phases import combine_global_phases_pass, CombineGlobalPhasesPass
+from .diagonalize_measurements import (
     DiagonalizeFinalMeasurementsPass,
-    iterative_cancel_inverses_pass,
-    IterativeCancelInversesPass,
-    measurements_from_samples_pass,
-    MeasurementsFromSamplesPass,
-    merge_rotations_pass,
-    MergeRotationsPass,
+    diagonalize_final_measurements_pass,
 )
+from .measurements_from_samples import (
+    MeasurementsFromSamplesPass,
+    measurements_from_samples_pass,
+)
+from .merge_rotations import MergeRotationsPass, merge_rotations_pass
 
 
 __all__ = [
-    # Quantum
     "combine_global_phases_pass",
     "CombineGlobalPhasesPass",
     "diagonalize_final_measurements_pass",
@@ -48,11 +36,4 @@ __all__ = [
     "MeasurementsFromSamplesPass",
     "merge_rotations_pass",
     "MergeRotationsPass",
-    # MBQC
-    "convert_to_mbqc_formalism_pass",
-    "ConvertToMBQCFormalismPass",
-    "decompose_graph_state_pass",
-    "DecomposeGraphStatePass",
-    "null_decompose_graph_state_pass",
-    "NullDecomposeGraphStatePass",
 ]
