+++ pennylane/compiler/python_compiler/transforms/quantum/merge_rotations.py
@@ -22,8 +22,8 @@ from xdsl.dialects import arith, builtin, func
 from xdsl.ir import Operation
 from xdsl.rewriter import InsertPoint
 
-from ..dialects.quantum import CustomOp
-from .api import compiler_transform
+from ...dialects.quantum import CustomOp
+from ...pass_api import compiler_transform
 
 # Can handle all composible rotations except Rot... for now
 composable_rotations = [
