+++ pennylane/compiler/python_compiler/visualization/collector.py
@@ -15,27 +15,27 @@
 which collects and maps PennyLane operations and measurements from xDSL."""
 
 from functools import singledispatchmethod
-from typing import Any, Union
+from typing import Any
 
 from xdsl.dialects import builtin, func
 from xdsl.ir import SSAValue
 
-from pennylane.compiler.python_compiler.dialects.quantum import AllocOp as AllocOpPL
-from pennylane.compiler.python_compiler.dialects.quantum import (
+from pennylane.measurements import MeasurementProcess
+from pennylane.operation import Operator
+
+from ..dialects.quantum import AllocOp as AllocOpPL
+from ..dialects.quantum import (
     CustomOp,
     ExpvalOp,
 )
-from pennylane.compiler.python_compiler.dialects.quantum import ExtractOp as ExtractOpPL
-from pennylane.compiler.python_compiler.dialects.quantum import (
+from ..dialects.quantum import ExtractOp as ExtractOpPL
+from ..dialects.quantum import (
     GlobalPhaseOp,
     ProbsOp,
     SampleOp,
     StateOp,
     VarianceOp,
 )
-from pennylane.measurements import MeasurementProcess
-from pennylane.operation import Operator
-
 from .xdsl_conversion import (
     dispatch_wires_extract,
     xdsl_to_qml_meas,
@@ -57,7 +57,7 @@ class QMLCollector:
         self.quantum_register: SSAValue | None = None
 
     @singledispatchmethod
-    def handle(self, _: Any) -> Union[Operator, MeasurementProcess, None]:
+    def handle(self, _: Any) -> Operator | MeasurementProcess | None:
         """Default handler for unsupported operations. If the operation is not recognized, return None."""
         return None
