+++ tests/python_compiler/conftest.py
@@ -30,11 +30,8 @@ try:
     from xdsl.passes import PassPipeline
     from xdsl.printer import Printer
 
-    from pennylane.compiler.python_compiler import Compiler
-    from pennylane.compiler.python_compiler.jax_utils import (
-        QuantumParser,
-        parse_generic_to_xdsl_module,
-    )
+    from pennylane.compiler.python_compiler import Compiler, QuantumParser
+    from pennylane.compiler.python_compiler.jax_utils import parse_generic_to_xdsl_module
 except (ImportError, ModuleNotFoundError):
     deps_available = False
